+++ bin/replicante/src/config/mod.rs
@@ -8,8 +8,6 @@ use serde_derive::Serialize;
 
 use replicante_logging::Config as LoggingConfig;
 use replicante_logging::LoggingLevel;
-use replicante_models_core::scope::Namespace;
-use replicante_models_core::scope::NsHttpsTransport;
 use replicante_service_coordinator::Config as CoordinatorConfig;
 use replicante_service_tasks::Config as TasksConfig;
 use replicante_stream::StreamConfig;
@@ -136,34 +134,4 @@ impl Config {
         Config::from_reader(include_str!("mock_config.yaml").as_bytes())
             .expect("mock config to load")
     }
-}
-
-/// Settings that will move to the DB once namespaces are fully introduced.
-#[derive(Clone, Default, Eq, PartialEq, Hash, Debug, Serialize, Deserialize)]
-pub struct TmpNsSettings {
-    /// HTTPS configuration for agent transport.
-    #[serde(default)]
-    pub https_transport: TmpNsSettingsHttps,
-}
-
-impl From<TmpNsSettings> for Namespace {
-    fn from(config: TmpNsSettings) -> Namespace {
-        Namespace {
-            ns_id: "tmp_global_namespace".into(),
-            https_transport: NsHttpsTransport {
-                ca_bundle: config.https_transport.ca_bundle,
-                client_key_id: config.https_transport.client_key,
-            },
-        }
-    }
-}
-
-/// HTTPS configuration for agent transport.
-#[derive(Clone, Default, Eq, PartialEq, Hash, Debug, Serialize, Deserialize)]
-pub struct TmpNsSettingsHttps {
-    /// PEM formatted bundle of CA certificates to validate agent certificates.
-    pub ca_bundle: Option<String>,
-
-    /// Path to a PEM formatted HTTPS client **private** key.
-    pub client_key: Option<String>,
 }